requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-amazon==8.12.0
pandas
boto3
requests
pyspark